fixed routes/fish.go
